LayoutTests/imported/w3c/web-platform-tests/css/css-writing-modes/text-combine-upright-parsing-digits-002-expected.txt


FAIL Computed value for `text-combine-upright: digits calc(1 + 1)` is `digits 2` assert_equals: expected "digits 2" but got "none"
FAIL Computed value for `text-combine-upright: digits calc(1 + 1 + 1)` is `digits 3` assert_equals: expected "digits 3" but got "none"
FAIL Computed value for `text-combine-upright: digits calc(2 + 2)` is `digits 4` assert_equals: expected "digits 4" but got "none"
FAIL Computed value for `text-combine-upright: digits calc(4 - 2)` is `digits 2` assert_equals: expected "digits 2" but got "none"
FAIL Computed value for `text-combine-upright: digits calc(6 - 3)` is `digits 3` assert_equals: expected "digits 3" but got "none"
FAIL Computed value for `text-combine-upright: digits calc(7 - 1 - 2)` is `digits 4` assert_equals: expected "digits 4" but got "none"
FAIL Computed value for `text-combine-upright: digits calc(2 * 1)` is `digits 2` assert_equals: expected "digits 2" but got "none"
FAIL Computed value for `text-combine-upright: digits calc(3 * 1)` is `digits 3` assert_equals: expected "digits 3" but got "none"
FAIL Computed value for `text-combine-upright: digits calc(2 * 2 * 1)` is `digits 4` assert_equals: expected "digits 4" but got "none"
FAIL Computed value for `text-combine-upright: digits calc(4 / 2)` is `digits 2` assert_equals: expected "digits 2" but got "none"
FAIL Computed value for `text-combine-upright: digits calc(6 / 2)` is `digits 3` assert_equals: expected "digits 3" but got "none"
FAIL Computed value for `text-combine-upright: digits calc(24 / 3 / 2)` is `digits 4` assert_equals: expected "digits 4" but got "none"
